backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
celery==5.3.4
flower==2.0.1
boto3==1.34.0
python-magic==0.4.27
Pillow==10.1.0
PyPDF2==3.0.1
PyMuPDF==1.23.14
reportlab==4.0.7
tesseract==0.1.3
opencv-python==4.8.1.78
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.6.0
